Add the Network.SPDY.Deserialize module, with a parser for raw frames.
This adds an attoparsec parser for the 'RawFrame' type. At the moment,
I'm planning to add a conversion function from 'RawFrame' to 'Frame',
rather than parsing frames directly from the input stream. I'm not
certain yet whether that would impose an unacceptable performance
penalty, but I think it could be useful. For example, a client or
server could choose to log the raw frames, showing where it thinks
they begin and end, as a debugging aid.