Add the Network.SPDY.Serialize module, with a serializer for raw frames.
The serializer uses the blaze-builder package; at the moment, there is
only a function for converting a single frame into a strict
ByteString. The serializer should be the inverse of the parser, so in
the next commit we can write some QuickCheck tests that exercise this
property.

One thing that's not quite clear is whether it's better to make a
single Write instead of a Builder, and use the 'fromWrite' function
once for the frame. This implementation uses Builders everywhere.  The
'writeByteString' function does write a strict byte string to a
buffer, so we can turn the payload into a Write, and thus use Writes
instead of Builders. Perhaps after the unit tests we can do a
Criterion benchmark to check whether one is better than the other.