Unit tests: check deserializing serialized raw frames.
Here's the first simple QuickCheck test: given an arbitrary raw frame,
check that deserializing the serialization of that frame gives the
same frame back.

The generators here are willing to generate garbage; in particular, we
don't care if the control frame types are meaningful, or that the
payload makes sense for legal control frame types. The only internal
constraint is that the payload length really is the length of the
payload.

The tests are integrated into the cabal build as a 'test-suite' of
type 'exitcode-stdio-1.0'.

[Aside: the changes in the cabal file include some whitespace-only
changes that I didn't mean to include, the result of running
'whitespace-cleanup' in emacs. However, I get 'corrupt patch' errors
from 'git citool' when I try to revise the patch by excluding those
lines. Ordinarily, I'd amend the patch, but I'm not going to try
*that* hard for just a few lines. Hmm. I wonder if the line
termination style changed, and that's the problem.]