Use test-framework to create the test program.
There are a couple of nice benefits to using test-framework along with
QuickCheck:

- we can give meaningful names to tests and groups of tests, and the
  program will print the names out while the tests execute. When there
  is a test failure, it's much easier to see which test failed.

- the test program (created by the 'defaultMain' function) will terminate
  with a non-zero exit code if any of the tests fail. This integrates
  nicely with cabal's test suite support -- at least the exitcode-stdio
  driver.

Interestingly, I'm now finding that the serialize-deserialize tests for
raw frames fail quite reliably, so I obviously didn't check as well as
I thought I did in preparation for the last commit. That's the next thing
to fix.