Rework how flags are represented in processed frames.
Using lists to represent sets of flags seems a little silly. The use
cases for flags are:

- testing whether they are set when reading frames

- setting them when writing frames

Using a list makes the first needlessly inefficient in time and space.

The repeated XUnknown flags are also bothersome. There really ought
to be just one way to capture flags we don't understand, say as a Word8
value, if it's indeed necessary at all.

I've reworked the collections of flags so that they are just newtype
wrappers around Word8, and I've added a two-parameter 'Flag' type class
so we can pair these newtypes with corresponding flag enumerations. Now,
flags are stored and queried more efficiently. For the time being, I've
forgotten about tracking whether there are flags we don't understand.

The new typeclasses are implemented in a separate Flags module, since
it seemed like a nice bit of self-contained functionality.