Add the skeleton code for converting frames to raw frames.
This adds to the 'Serialize' module a skeleton of the code to convert
frames into raw frames. The exported functions are

 toRawFrame :: Frame -> RawFrame

and

 frameToByteString :: Frame -> ByteString

which uses 'toRawFrame' and 'rawFrameToByteString'. 'toRawFrame' is
by far the most complex, and splits the task into

 - serializing the header

 - serializing the flags, if any, and the payload length

 - serializing the payload

The complexity will mostly be in serializing the payload for control
frames, where we'll need to use Builders to encode the details of
each different type of control frame.