Add the ToBuilder type class, and some trivial instances.
Most of a ControlFrameDetails value becomes payload when converting
to a raw frame. In order to write the conversion, we'll use blaze-
builder to build a payload ByteString. In order to avoid having to
pollute the 'Serialize' module with a bunch of 'XBuilder' functions,
I've added another trivial type class, ToBuilder, and the most trivial
instances. I'll add more instances over the next few commits.