Add a 'ToBuilder' instance for 'HeaderBlock'.
This gets us partway to serializing header blocks. For the rest, we
need to be able to do zlib compression using a custom compression
dictionary; unfortunately, it looks like we have to do that in the IO
monad, and so the next step will be to make some type signature changes
that allow this.