Support header block compression when serializing control frame payloads.
The spec says that the headers transmitted in frames like SYN_STREAM
are supposed to be compressed, with the following restrictions:

- there is a prescribed compression dictionary

- there is a single compression stream for all the headers going in one
  direction on a connection (and there is one connection per domain). This
  amortizes the cost of the compression context. The stream has to be
  flushed at the end of each header block.

This requires two changes in the types for the 'toRawFrame' function and
some of its associated functions:

1. the functions must return their results in the IO monad; we're using
   the zlib-bindings package for access to a low-enough zlib interface
   that we can control the compression dictionary and when the compression
   stream is flushed. We need to stay in IO in order to ensure that the
   operations on the 'Deflate' value, which represents the compression
   stream, are properly sequenced.

2. the functions need to receive a 'Deflate' argument.

I considered layering 'StateT Deflate' on top of IO, but I'm not sure that
it's worth the trouble.

I've implemented compression for header blocks and made the necessary
type changes in the process. I've also completed the handling of
SYN_STREAM frames in 'toControlPayloadBuilder'.

The logical next step is to implement the conversion from raw SYN_STREAM
frames into processed frames, so we can add some unit tests that convert
a processed frame into a raw frame and back. That's a weak test compared
to interoperating with an existing server implementation, but that will
have to wait for a little while.