Add a test for converting DATA frames to raw frames and back.
This test is similar in spirit to the deserialize-serialize test for
raw frames, except that we test that we can serialize a DataFrame
variant of Frame to a raw frame, and then convert back to an equal
DataFrame.

Currently, we limit the tests to DataFrame frames only because the
Arbitrary instance generates only that variant. Most of the work was
in setting up the test in the new FrameTests module, and adding
Arbitrary instances in the Instances module. In the next commit, I'll
enhance the Arbitrary instance for Frame so it also generates
ControlFrame values, but I'll limit the required Arbitrary instance
for ControlFrameDetails so that it generates only the SynStream
variant.

Sadly, I've had to resort to using 'morallyDubiousIOProperty' to
implement the 'prop_toRawFrameAndBack' property, since the conversion
to raw frames and back is in the IO monad. I think that at one time
QuickCheck had a way to test properties that do IO, but QuickCheck2
seems to have this only reluctantly. I'll see if I can find a better
way to do this kind of test.