Split out an Internal.Deserialize module.
In some tests that I have yet to check in, I've found a problem with
either the conversion of processed SYN_STREAM frames into raw frames,
or the conversion from processed frames to SynStream control frames.
Unfortunately, there is a fair bit of supporting code involved, in
particular ToBuilder definitions and attoparsec parsers, all of
which is hidden in the Serialize and Deserialize modules. I'd like
to unit test that code, I've decided to pull that code into special
Internal.Serialize and Internal.Deserialize modules which export
everything, but aren't exposed in the cabal package.

This commit adds the Internal.Deserialize module. In the next commit,
I'll add Internal.Serialize.