Add the Compression module.
Stuff related to compression should really be in a single module.
I've added a Compression module and moved 'compressionDictionary'
from the Frames module, where it's a little out of place.

In the next couple of commits I'll move the compression and
decompression functions into this module too, and remove a little
bit of repeated code.