Factor out common code in 'compress' and 'decompress', and fix a bug.
It bugged me that 'compress' and 'decompress' had almost identical
implementations, so I decided to eliminate the repeated code with
a helper function. This turned out to be fortuitous, because I
found a bug: I was discarding the result of calling 'flushInflate'!

'flushInflate' has the type

  Inflate -> IO ByteString

whereas 'flushDeflate' has the type

  Deflate -> (IO (Maybe ByteString) -> IO a) -> IO a

In 'compress', I was ignoring the () result of 'flushDeflate', and
when I implemented 'decompress' I didn't realize that 'flushInflate'
takes no popper because it returns a final fragment of output,
and I discarded the output of 'flushInflate'. For small amounts of
input data, this is probably all the decompressed output!

This explains why the to-raw-frame-and-back test was failing when
I allowed the Frame generator to produce SYN_STREAM frames (yet to be
checked in): when decompressing the headers, most of the data was
thrown away. With this change, those tests now pass, so I'll add the
generator changes in the next commit.

This would have taken a little while to figure out even with unit tests
targeted at header compression and decompression, or even just byte
string compression and decompression. Trying to eliminate the
repeated code forced me to look at the types more closely. I'll still
add the tests for the supporting code though.