Add the credential slot field for SYN_STREAM frames.
This is the first step in updating the frame definitions so they are
consistent with the February 2012 draft of the SPDY spec. I was going
to work on conversions for more control frames to raw frames, but I'm
starting to encounter more deviations from the spec, so it's time to
spend a few commits catching up. This will be the first of several
such commits, and at the end I'll update the Frames module comment
to refer to the correct spec version.