Add a CREDENTIAL control frame.
The current draft spec for SPDY 3 adds a CREDENTIAL control frame,
used by a client to insert a credential into the server's credential
vector. This commit simply adds the necessary variant to the
ControlFrameDetails type, and makes a few associated minor changes.
The low-level serialization and deserialization code will come in a
future commit.

Note that both the SYN_STREAM and CREDENTIAL frames have 'slot' fields,
but they are different sizes. This is inconsistent, but from what I've
been able to gather in the spdy-dev Google group, the inconsistency
isn't a mistake. Although the CREDENTIAL frame was proposed with a
16-bit Slot field, in SYN_STREAM the Slot field occupies what was a
reserved but unused byte in the SPDY 2 version of SYN_STREAM. For now,
it looks like the inconsistency will simply stick.

It would certainly make sense to have one Slot type, eight bits wide,
and simply use a byte of padding when serializing and deserializing,
since it doesn't really make sense for the server to support a
credential vector with slots that cannot be referenced at stream
creation time. For the moment, I've declared a 'Slot16' type instead,
just to be faithful to the spec; I may change my mind about that.