Simplify the way we deal with Flags.
Thinking about it a bit, I realized that the Flags module was more
complicated than it needed to be. Instead of Flag and Flags type
classes, we needed only the Flag type class and a single Flags
data type with a phantom type parameter.

This allows us to get rid of the different XFlags container types
in the Frames module, together with their ToWord8 instances, and
allows us to avoid some tedious repeated code in the unit test
Instances module.