Support converting RST_STREAM frames to and from raw frames.
The bulk of the work here was implementing the builder and the
parser for the TerminationStatus type. In order to do this, there
are 'tsX' Word32 constants defined in the Frames module for each
of the expected TerminationStatus variants. The builder and parser
for the RstStream content are then trivial.

I've added a build-parse unit test for TerminationStatus, and
have enhanced the Arbitrary instance for ControlFrameDetails so
it generates the RstStream variant. The tests all pass.