Fix a bug in the Arbitrary instance for SettingID.
The SettingID type is almost an enumeration, except for the
SettingsOther variant, which is designed as a catch-all for setting
IDs that we don't understand.

When generating arbitrary SettingIDs, we need to make sure that for
the SettingsOther case, the Word32 that we generate isn't one of
the predefined constants associated with the other SettingID variants.
Otherwise, a build-parse test (or a to-raw-and-back test involving
a SETTINGS frame) can fail.

In fact, I had this bug. Since the reserved Word32 values were
consecutive, I restricted the Word32s generated for the SettingsOther
case to be >= (length fixedValues), where 'fixedValues' holds all
the contructor-only variants. But the reserved Word32 values start
at 1, not 0! As a result, we'd occasionally generate

  SettingsOther 8

which in fact conflicts with the reserved Word32 value for

  SettingsClientCertificateVectorSize

The 100 tests for the build-parse test for SettingID would
eventually fail after a few tries, and so would the to-raw-and-back
test for Frames, but you could also run them several times without
a failure. I've run them more than 100 times now without any failures,
so I think I've caught the problem.