Support converting WINDOW_UPDATE frames to raw frames and back.
We needed a builder and a parser for DeltaWindowSize, but these
were trivial so I added them and a build-parse unit test. Implementing
the conversion for WINDOW_UPDATE frames was then easy, and I added
a case for WindowUpdate in the Arbitrary instance for
ControlFrameDetails. I've run the unit tests successfully for 100
iterations.