Support converting CREDENTIAL frames to raw frames and back.
I had already accidentally added the builder for the Credential
variant of ControlFrameDetails in the last commit, so there was
just the parser to finish up, and another case to add in the
Arbitrary instance for ControlFrameDetails. I've run the tests for
100 iterations successfully.

CREDENTIAL is the last of the control frames, so at this point the
serialization and deserialization of frames is finally done. It's
time to do a little cleanup and then to start thinking about how to
structure the client and the server code. I think it might be best to
start with a stand-alone client, as long as I can get running a SPDY 3
server of some kind.