Move several numeric constants from Frames into their own module.
The Frames module defines several numeric constants whose only use
is in serializing and parsing SPDY frames. On the one hand, it's nice
to have these constants in the Frames module, since they are close
to the types with which they are associated, and that can be a help
when reading the source and the spec at the same time -- it's easy
to confirm that they are correct. However, the constants don't really
have a use to clients of this package, so I've decided to move them
into their own internal module, Internal.Constants.

Everything still builds, and the unit tests still pass.