Use 1000 iterations for the to-raw-and-back test.
The to-raw-and-back test has a lot of ground to cover, since there
are several different frame types. The default of 100 tests seems
a little skimpy. After poking around for a bit, I've found how to
change the default: using test-framework's 'plusTestOptions' function.

The only oddity was figuring out how to use the TestOptions type;
what was confusing and at first a little annoying is the lack of a
set of default TestOptions. However, TestOptions is just a type
synonym for

  TestOptions Maybe

where each of the fields in

  TestOptions f

is wrapped with an 'f'. Further, TestOptions = TestOptions' Maybe
is an instance of Monoid, where 'mempty' is a TestOptions value
with Nothing in every field. You can override the number of
generated tests by using

 mempty { topt_maximum_generated_tests = Just 1000 }

With this change, the to-raw-frame-and-back test now runs for 1000
iterations instead of the default 100. I've run the full set of tests
for over 100 times without any failures.