Add a set of rough notes on SPDY client considerations.
The next major thing to work on is a SPDY client. I'd like to make
one that can talk to the 'flip' server that resides in the Chromium
source tree, perhaps without using TLS to start off.

However, the spec is shot through with various rules on how a client
and server are supposed to interact, so I've reread the spec and
taken some rough notes on things that directly concern the client.

Out of this, I've gotten a rough feel for what state a SPDY client
needs to maintain. I think it's enough to start iterating on the
code a little bit, though I might try sketching out a rough finite
state automaton before I begin banging out any code. I doubt this
will see too many edits, but I'll keep it here for now.