Stub a SPDY client with a partial ability to handle PING frames.
There are a lot of rough edges here, but with this code we can PING a
modified 'flip' server, and get a response. But several things are
missing:

- if we send a PING and the response never comes, we don't time
  out. This means that a 'ping' client API call can hang forever if
  there is a network outage, or if the server does not support
  pings. (and see below)

- we respond properly to PINGs that are responses to pings we have
  initiated, but we don't respond properly to PINGs initiated by the
  server

- we don't properly detect when the other side seems to have
  gracelessly aborted the connection, and cleanly shut down

- a few other TODOs noted in the code.

In order to manually test this code, I had to modify the 'flip' server
in the current chromium source tree in two ways:

- change the hard-coded SPDY version to 3 from 2

- implement support for responding to PING frames in the server

It turns out that the flip server is not a complete server
implementation. In addition to ignoring PING frames, it also seems to
ignore WINDOW_UPDATE and so cannot do flow control. It's very
bare-bones. In order to try this stuff out, I had to hack the flip
server to accept SPDY 3 connections, and to respond to PING
frames. This was straightforward, and the little sample 'ping'
application works with it. At least this confirms that we can read
PING frames sent by a Google SPDY server; as it turns out, we can also
read a simple SETTINGS frame, one that is automatically sent by the
flip server whenever it accepts a connection.

Unfortunately, the flip server's limitations mean that we may need to
create FFI bindings for a hacked TLS library with support for the NPN
extension a little sooner than I was hoping. It's either that, or
modify the flip server to support more of the SPDY 3 spec.