Time out if we don't receive a response to a PING frame.
The current half-witted implementation will hang forever if the
server does not properly support PING frames, or if there is a
network error and we don't receive a response. Use a time out
and indicate in the result whether we received a response or we
timed out waiting.

The timeout is given in a PingOptions value, and there is a new
two-variant PingResult type for the result. Since the actions
to take in the time-out and response cases involved similar
time arithmetic, I extracted the 'millisSince' function and did
a little more refactoring in the 'ping' function.

I've tested this manually by commenting out the normal response
to received PING frames in 'readFrames', so that we never write
the time of the response into the MVar, and observed that we
time out as expected.