Add a simple 'priority channel' module, Internal.PriorityChan.
In order to respond to server-initiated PING frames, we're supposed
to send a matching PING frame as soon as possible, sending it before
DATA frames or other control frames that we might have pending. One
straightforward way is to have a thread for sending frames (or,
more generally, 'taking actions'), which receives jobs to do on a
priority queue. PING frames would be given immediate priority.

Since the producer and consumer will be in different threads, and
since we don't want the consumer to busy-wait when there is nothing
to do, we have to do something a little fancier than have a priority
queue in an IORef. I've added a simple module implementing a
'priority channel'. This is a priority queue in an IORef along with
a semaphore that has credits only as along as the queue is non-empty.
The priority queue is the one provided in the 'fingertree' package.

This implementation does not limit the size of the priority queue,
so if the producer generates jobs very quickly it could strain
memory resources. In practice, I don't think this is very likely,
so I'll run with this implementation for now.

The new module isn't used anywhere yet. I'll add some unit tests
first, and then use it to add PING responses to the Client module.