Use a priority channel for 'outgoing' jobs.
Instead of directly writing frames to the outgoing socket handle,
or flushing the handle, send jobs to a priority channel. A single
worker thread receives jobs and performs them. We can prioritize
jobs according to a SPDY stream priority, or using a special 'ASAP'
override priority.

This change adds no new functionality, but removes the 'writeFrame'
and 'flushConnection' functions in preference for something a
little fancier. In the next commit or so, we'll exploit that to
respond to server-initiated PING frames with an ASAP priority,
so we send the response PING ahead of other data or control frames.

The little 'ping' test application still works with the flip server,
though I've noticed a slight bump in the round-trip times when talking
on the loopback interface. Now pings that tended to take around a
millisecond more often take two or so milliseconds. It could be that
the priority channel imposes a significant latency. We'll see if it
becomes a problem when we're really transmitting data.