Respond to server-initiated PING frames.
Our PING frame handling isn't entirely correct. Our current handling
applies only to client-initiated pings, but not for server-initiated
ones. In the server-initiated case, we respond with an identical PING
frame as soon as we can.

I've added functions 'isServerInitiated' and 'isClientInitiated' to
the Frames module, and refined the handling of PING frames in the
Client module.

This code has not seen any testing, since we don't have a server that
generates pings but doesn't require TLS and the NPN extension. We may
need to start building a stub server now. In fact, that might be a
good idea, especially if we can write some unit tests that exercise
client-server interaction. This would be a good way to check some
things we might not encounter all that often in practice.