Extract 'controlFrame' and 'pingFrame' functions in the Client module.
We'll soon have reason to create different types of control frames,
so a 'controlFrame' function that inserts the protocol version for
a connection seems like a good idea. Similarly, I've extracted a
'pingFrame' function from 'ping'.

A bunch of this stuff will eventually have to be shared between the
client and the server implementation. I'm starting to think that it
would be a good idea to start implementing a server soon, to help
figure out which parts of the code should be common. There is quite
a bit that should be.