Allow queuing GOAWAY frames.
In order to implement connection shutdown, we'll need to be able
to send GOAWAY frames. To do that, we need to record the last stream
ID for which we sent a SYN_REPLY or RST_STREAM frame. (Aside: I'm
not really sure whether all RST_STREAM frames really count, or just
ones that are immediate replies to SYN_STREAM frames. For now, I've
assumed that it's OK to treat all RST_STREAM frames the same.)

This commit adds to Connection an IORef containing the last such
stream ID. I've modified the 'doOutgoingJobs' function to set the
stream ID whenever it sends a SYN_REPLY or RST_STREAM frame, and
I've added a function 'goAwayFrame' for creating GOAWAY frames. We'll
be using that in the next couple of commits.