Add the ability to shut down a connection.
One obvious deficiency with the client is that there is no way to
gracefully shut down a connection. The modified flip server does
not seem to perform a proper shut down by sending a GOAWAY frame,
but instead just closes the network connection. On our end, the
frame reading thread reads an empty byte string off the socket
handle, and the result is a parse error. However, we log the error
and then continue merrily along.

In order to cleanly shut down the connection, there are a few things
that we need to do:

- the connection needs to be removed from the client's connection
  map; in order to help with this, I've added the new 'connClient'
  and 'connKeys' fields to the Connection type. Although we don't
  currently try to determine if two distinct connection keys are
  equivalent, we will eventually, so 'connKeys' is an IORef to
  a list of ConnectionKeys

- we may need to send a GOAWAY frame, but we don't want to do it
  more than once. To help with this, I've added the
  ConnectionLifeCycleState type and the 'connLifeCycleState' field.
  The Open variant of ConnectionLifeCycleState carries a timestamp
  for the most recent activity on the connection, so that we can
  implement an idle connection timeout at some point. This timeout
  is bumped by the 'touch' function whenever we read or write a
  frame.

- the 'closeConnection' function takes care of updating the
  life cycle state, sending a GOAWAY frame if necessary, and
  asking the outgoing job thread to stop. The action carried
  by the Stop job closes the network connection and changes
  the lifecycle state to Closed.

I'm not sure that I like having the outgoing jobs thread be
responsible for closing the network connection, but it seems
like the easy way to do it.

The final change is in 'readFrames'. Now, if we encounter an
error when parsing frames, we shut down the connection, sending
a GOAWAY frame with the 'protocol error' status. Right now our
only source of non-exception errors seems to be when there is
a parse error parsing a frame, so, assuming that there are no
bugs in the parser, signalling a protocol error seems like a
good choice.

One glaring omission here is that we do not defend against
exceptions in the frame reading thread or the outgoing job thread,
so an error could kill one of those threads. In particular, if
the outgoing job thread is killed, we will not close the socket
handle and consequently leak a file descriptor. I figured that is
best fixed in a separate commit.