Stub out an 'initiateStream' function in the Client.
The next step is to download files from the 'flip' server; in order
to do that we'll need to be able to initiate a stream with a given
set of headers and some optional data, and accept response headers
and data. This change just stubs out the function declaration, and
adds a dependency on the 'enumerator' package.

At first I thought I would try to avoid any dependency on fancy
things like iteratees or conduits, but I think that picking one of
these libraries will probably simplify the types a bit, so I'll
try 'enumerator' first.