Add an Error module, with a SPDYException type.
This is rather long overdue, but we need a way to express SPDY-specific
errors. I think I'll take a page out of Michael Snoyman's book for
'zlib-bindings', and declare a SPDY exception type. This type will
doubtless expand with more variants, but these are two that we will
be able to use in the short term.