Add a 'nextStreamID' function in Client.
In order to initiate streams, we'll need to be able to allocate stream
IDs. This is very similar to ping IDs, except that stream IDs aren't
allowed to wrap, so we throw an 'OutOfStreamIDs' exception when
they do. This function is still internal to the module, but I'll use
it shortly.