Add a 'synStreamFrame' function to the Client module.
To initiate a stream, we'll also need to create and send a
SYN_STREAM frame. This commit adds a utility function to create
the frame.

I've also added a default priority value in Client, and a 'noSlot'
value in the Frames module. I don't have any real basis for using
a default priority of 4, except that it's close to the middle and
allows bumping priorities a couple of levels.

Although 'synStreamFrame' calls 'nextStreamID' and can fail with
a SPDYException, I haven't put a handler in that function, since it
isn't exposed in the API. I think it would be a good idea to create
a special 'bracket' function that is used in all the API calls; it
might also be used in the reader thread and outgoing job threads,
though I'm not sure that the exception handling would be the same
there. Those might need to be one-of handlers.