Add a Stream data type to the Client.
We now track some fundamental things about streams, including

- the ID and priority
- the source for data destined to the remote endpoint
- the consumer for headers from the remote endpoint
- the consumer for data from the remote endpoint

None of this is used yet. I will probably change the type of
'initiateStream' so that it doesn't use enumerators and iteratees,
opting for the same approach used in Stream.

The Connection now carries an IORef with a map from StreamIDs to
Streams, and sports a couple of utility functions for adding,
removing, and looking up streams, and functions for submitting
headers and data from the remote endpoint. In order to add this
stuff, I also derived the default instance of Ord for StreamID.

In the next few commits, I'll add handlers in the reader thread
for SYN_REPLY, HEADERS, and DATA frames, and then finish off the
implementation of 'initiateStream'. At that point, we should be
able to write a little program that downloads a file from the flip
server, and prints the headers and data on the standard output.