On second thought, don't use 'enumerator' types in 'initiateStream'.
I could be wrong, but I don't think it's possible to get just the next
element from an enumerator, leaving it in a state where it can be
resumed. This means that the third argument of 'initiateStream' shouldn't
be an Enumerator.

It should be possible to just submit the next element to an iteratee,
but there is no guarantee that the resulting iteratee will be able to
consume more input. While we could retain the Iteratee types in the
last two arguments of 'intiateStream', I've decided not to.

Besides, now the data source and the header and data consumers can
simply be stored in the newly-created Stream. This should be enough
to get something that works; if the types turn out not to be ideal,
it should be relatively easy to change them later.

Since 'intiateStream' was the only thing that used Iteratee or Enumerator,
I've removed the dependency on the 'enumerator' package in the cabal file.