Implement just the basics of down-loading a data stream from the server.
This commit fills out enough of 'initiateStream', and adds enough
frame handlers in 'readFrames', that we can can open a half-open
stream in order to download the contents of a url from the 'flip'
server. I've tried this out and confirmed that it works with a small
test program, which I'll add in the next commit.

There is still a fair bit of stuff that is missing here. In particular:

- other than the first chunk of data, there is no way to submit further
  further data chunks to the remote endpoint

- if the list of headers is too big, we have no way to detect it, and
  have no way to continue the headers in additional HEADERS frames

- we do not respond to RST_STREAM frames, so if the server signals an
  error we will not properly clean up the stream

- we have no way to cancel open streams when there is a connection
  error, or when we receive a GOAWAY

- we do not defensively handle exceptions that can occur when calling
  into the application-supplied data source, header consumer, or
  data consumer

But, despite the short-comings, it seems we can correctly receive
SYN_REPLY and DATA frames.

In the process of making these changes, I realized that the proper
application-facing interface for 'initiateStream' should use the
client and the connection key. However, now there is an
inconsistency between 'ping' and 'initiateStream', in two respects:

- 'ping' puts the client and connection key arguments last, but
  'initiateStream' puts them first

- 'ping' collects all the ping-specific arguments in a value of type
  PingOptions, so there can be some reasonable defaults, but
  'initiateStream' does not. We should pick a style and stick with it.

I'll resolve these issues in future commits.

With respect to the issue of not sending all the data that might be
produced by the data-producer action, there isn't a lot of testing
that we can easily do here with the flip server -- or at least I don't
think there is. Maybe if we can arrange for some responses to HTTP
POSTS we can try the uploading functionality, but I'm not sure how
much we can really exercise it.

However, were we to start implementing a server, we could write some
unit tests that exercise the client and server talking to one
another. There is quite a bit of code that could be factored out into
a common 'EndPoint' module, so perhaps we should spend some time on
that. I've also noticed that it takes quite a while to download a
large file like '/usr/share/dict/words', so it would also be a good
idea to do a little profiling and see where the performance problem
lies.