Add SGet.hs, a small program for fetching a resource from the flip server.
SGet.hs is slightly more complicated than SPing, and downloads a resource
from the flip server. Currently, it prints the headers and data on stdout,
and logs all frames on stderr, so to make the output useful it's best
to separate the two streams with redirection, e.g.

 ./sget host port file > file 2> file.log

will put the output in <file> and the logged frames in <file.log>.

I've tried it out with a handful of files I've got loaded up in the
modified 'flip' server, including the contents of '/usr/share/dict/words'
on my system, and it can recover the data faithfully.

I've also added a Makefile that can build the 'sping' and 'sget' programs,
and I've updated '.gitignore' to ignore them and the temporary directory
used for the build output.