Produce a version of 'sget' with profiling enabled.
This adds a target for an 'sget-prof' binary for SGet, with GHC time and
space profiling enabled.

The profile for 'sget-prof' when transferring the 'words' file isn't
terribly elluminating; it says that almost all the time is spent in
'readFrames', which is to be expected. However, none of the functions
called by 'readFrames' appears to be an obvious culprit.

Using the unix 'time' program to time the execution of 'sget' for this
file indicates that more than half the time is spent doing system calls:

  $ time ./sget 127.0.0.1 10041 words > words 2> words.log

  real                      0m6.137s
  user                      0m2.540s
  sys                       0m3.572s

which suggests that block-buffering stdout and stderr might solve the
problem. I'll try that separately.