Block-buffer 'stdout' and 'stderr' in SPing too.
It turns out that a lot of the time that we're measuring in SPing is
often the time to log frames to 'stderr'. Block-buffering 'stderr' and,
for good measure, 'stdout' improves the performance of SPing too. Now
many of the measured round-trip times round to zero milliseconds when
we're talking to the flip server on the loopback interface, which seems
reasonable.