Send WINDOW_UPDATE frames from the client.
SPDY flow control requires that the client send WINDOW_UPDATE frames
when it consumes data sent by the server and frees up buffer space.
Once a DATA frame arrives with the FIN flag set, the client sends no
further WINDOW_UPDATE frames.

Since the client immediately passes received data chunks to the
data consumer associated with the stream, we need to change the type
of the consumer so it tells the client how much buffer space is
released. The consumer now has the type

  Maybe ByteString -> IO DeltaWindowSize

so it can say how many bytes have been released each time it is fed
more data.

It's possible that in some circumstances the application may take a
while to release buffer space, and so this interface may not be
sufficient. We should really add a function to the API specifically
for reporting changes in the window size when the data consumer
doesn't report window size changes. However, for many circumstances,
this interface may be enough. I'll make a separate commit to add
another API function.

These changes required a modest compensating change in the SGet test
program. There, the data consumer immediately writes bytes out to
stdout, so it always reports that the entire byte string is consumed.

The 'flip' server does not implement flow control, and in fact it has
absolutely no response to WINDOW_UPDATE frames. I've modified my version
of the server so that it logs WINDOW_UPDATE frames, but I currently
have no third-party server against which I can test my understanding of
flow control. That may need to wait until we can use TLS + NPN.