Don't send WINDOW_UPDATE frames if there is no change in window size.
It's entirely possible that the data consumer for a stream will report
a zero change in the window size. When this happens, there is no point
in sending a WINDOW_UPDATE frame to the sending endpoint. It's probably
harmless to do so, but why incur an unnecessary flush on the outgoing
network connection? I've guarded against this in 'readFrames'.