Add a public 'updateWindow' API function to the Client module.
There is a chance that application code will need to explicitly
participate in the SPDY flow control by signalling when there is
extra space in the window for data received from the remote endpoint.
I've added an 'updateWindow' function to do this, and factored out
the common code between it and 'readFrames' into "updateWindow'".

This isn't used yet, but I think it is necessary in the API.