Add a StreamOptions type and use it in 'initiateStream'.
The signature of 'initiateStream' is very long, and there may well be
some other arguments we'll need to add in the future. It is also
inconsistent with the signature for 'ping', which collects the ping-
specific arguments in a PingOptions type. So it seems reasonable to
add a StreamOptions type and a default 'defaultStreamOptions' value.

The 'defaultStreamOptions' value lets us get rid of 'defaultPriority',
and use simply Priority instead of (Maybe Priority) without burdening
application code. I've made the necessary adjustments in the SGet
module, and confirmed that I can still download files from the flip
server.

I've kept the list of headers out of StreamOptions, since the only
essential argument when creating a stream is the list of initial
headers for inclusion in the SYN_STREAM frame.

There is still one minor inconsistency between 'ping' and
'initiateStream'.  In 'ping', the client and connection key arguments
come after the options, whereas in 'initiateStream' they come before
the headers and options.  I'll pick one style in a subsequent commit.